+++ rtiow/C/a11-2.c
@@ -305,7 +305,7 @@ void make_camera(camera *cam, const vec3 lookfrom, const vec3 lookat,
 		const vec3 vup, float vfov, float aspect) {
 //	vprint(lookfrom);printf("\n");
 	vec3 u, v, w;
-	float theta = vfov*M_PI/180;
+	float theta = vfov*(float)M_PI/180;
 	float half_height = tanf(theta/2);
 	float half_width = aspect * half_height;
 
@@ -370,7 +370,7 @@ int main() {
 	int ny = 100;
 	int ns = 100;
 	printf("P3\n"); printf("%d %d\n", nx, ny); printf("255\n");
-	float R = cosf(M_PI/4);
+	float R = cosf((float)M_PI/4);
 	hittable_t world[] = {
 		HSTART,
 		HSPHERE(0, 0, -1, 0.5, MLAMBERTIAN(0.1, 0.2, 0.5)),
@@ -396,6 +396,7 @@ int main() {
 				float u = ((float)i + random_f()) / (float)nx;
 				float v = ((float)j + random_f()) / (float)ny;
 #ifdef DEBUG
+//				printf("u=%.6f v=%.6f\n", u, v);
 //				printf("u=%g v=%g rfcnt=%lu riuscnt=%lu\n", u, v, rfcnt, riuscnt);
 				vec3 uv = {u, v, 0};
 				printf("uv=");vprint(uv);printf(" \n");
