+++ rtiow/CPP/camera.h
@@ -16,7 +16,7 @@ public:
 		float aperture = 0;
 		float focus_dist = (lookfrom-lookat).length();
             lens_radius = aperture / 2;
-            float theta = vfov*M_PI/180;
+            float theta = vfov*(float)M_PI/180;
             float half_height = tanf(theta/2);
             float half_width = aspect * half_height;
             origin = lookfrom;
@@ -33,7 +33,7 @@ public:
         camera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov, float aspect,
                float aperture, float focus_dist) {
             lens_radius = aperture / 2;
-            float theta = vfov*M_PI/180;
+            float theta = vfov*(float)M_PI/180;
             float half_height = tanf(theta/2);
             float half_width = aspect * half_height;
             origin = lookfrom;
